Improve from_list logic and fix type hints For some reason, the from_list method was originally making the instance in which it then (optionally) checked for duplicate packets based on ID. However during the original parsing, packets were stored as dicts with IDs as keys, this means even if there were some duplicates, they'd just get overridden and we'd only pass the last packet in the list with that ID. This meant the duplicate check would never actually catch any duplicate packets.
This logic was also not ideal because it went through the list of
packets server times for no reason, when it could all be done in a
single loop. This is way more efficient which would become noticable if
the list of packets was long enough.

It also fixes an invalid type hint, which specified the packets argument
should be a List[Type[Packet]], however this would mean we expect the
list to contain the actual "Packet" class (or it's subtypes), when in
fact we want instances of this type. Not only that, but the function
can't even process any other type of packets than ClientBound and
ServerBound ones, so we shouldn't specifiy that it takes any arbitrary
packet, but rather a union of the two.